modified statistics to use a long for the sum of integers to avoid overflow

+++ src/main/java/com/deathrayresearch/outlier/columns/FloatColumn.java
@@ -89,7 +89,7 @@ public class FloatColumn extends AbstractColumn implements NumReduceUtils {
     return ColumnType.FLOAT;
   }
 
-  public float sum() {
+  public double sum() {
     return StatUtil.sum(this);
   }
